// dashboard/src/utils/types.tsx
export type Guild = {
  id: string;
  name: string;
  icon?: string;
  banner?: string;
  description?: string;
  owner?: boolean;
  permissions?: string;
};

export type MenuGroup = {
  name: string;
  path: string;
  routes: Route[];
};

export type Route = {
  name: string;
  path: string;
};

export type MenuItem = {
  name: string;
  value: string;
};